clean up comment deep link and user slugs
